+++ src/tightdb/table.cpp
@@ -631,6 +631,9 @@ void Table::do_insert_column(Descriptor& desc, size_t col_ndx, DataType type,
 {
     TIGHTDB_ASSERT(desc.is_attached());
 
+    if (TIGHTDB_UNLIKELY(name.size() > Descriptor::max_column_name_length))
+        throw LogicError(LogicError::column_name_too_long);
+
     typedef _impl::DescriptorFriend df;
     Table& root_table = df::get_root_table(desc);
     TIGHTDB_ASSERT(!root_table.has_shared_type());
@@ -2447,6 +2450,8 @@ StringData Table::get_string(size_t col_ndx, size_t ndx) const TIGHTDB_NOEXCEPT
 
 void Table::set_string(size_t col_ndx, size_t ndx, StringData value)
 {
+    if (TIGHTDB_UNLIKELY(value.size() > max_string_size))
+        throw LogicError(LogicError::string_too_big);
     if (TIGHTDB_UNLIKELY(!is_attached()))
         throw LogicError(LogicError::detached_accessor);
     if (TIGHTDB_UNLIKELY(ndx >= m_size))
@@ -2471,6 +2476,8 @@ void Table::set_string(size_t col_ndx, size_t ndx, StringData value)
 
 void Table::insert_string(size_t col_ndx, size_t ndx, StringData value)
 {
+    if (TIGHTDB_UNLIKELY(value.size() > max_string_size))
+        throw LogicError(LogicError::string_too_big);
     TIGHTDB_ASSERT(col_ndx < get_column_count());
     TIGHTDB_ASSERT(ndx <= m_size);
 
@@ -2503,6 +2510,8 @@ BinaryData Table::get_binary(size_t col_ndx, size_t ndx) const TIGHTDB_NOEXCEPT
 
 void Table::set_binary(size_t col_ndx, size_t ndx, BinaryData value)
 {
+    if (TIGHTDB_UNLIKELY(value.size() > max_binary_size))
+        throw LogicError(LogicError::binary_too_big);
     TIGHTDB_ASSERT(col_ndx < get_column_count());
     TIGHTDB_ASSERT(ndx < m_size);
     bump_version();
@@ -2518,6 +2527,8 @@ void Table::set_binary(size_t col_ndx, size_t ndx, BinaryData value)
 
 void Table::insert_binary(size_t col_ndx, size_t ndx, BinaryData value)
 {
+    if (TIGHTDB_UNLIKELY(value.size() > max_binary_size))
+        throw LogicError(LogicError::binary_too_big);
     TIGHTDB_ASSERT(col_ndx < get_column_count());
     TIGHTDB_ASSERT(ndx <= m_size);
 
@@ -2600,9 +2611,13 @@ void Table::set_mixed(size_t col_ndx, size_t ndx, Mixed value)
             column.set_double(ndx, value.get_double());
             break;
         case type_String:
+            if (TIGHTDB_UNLIKELY(value.get_string().size() > max_string_size))
+                throw LogicError(LogicError::string_too_big);
             column.set_string(ndx, value.get_string());
             break;
         case type_Binary:
+            if (TIGHTDB_UNLIKELY(value.get_binary().size() > max_binary_size))
+                throw LogicError(LogicError::binary_too_big);
             column.set_binary(ndx, value.get_binary());
             break;
         case type_Table:
@@ -2646,9 +2661,13 @@ void Table::insert_mixed(size_t col_ndx, size_t ndx, Mixed value)
             column.insert_double(ndx, value.get_double());
             break;
         case type_String:
+            if (TIGHTDB_UNLIKELY(value.get_string().size() > max_string_size))
+                throw LogicError(LogicError::string_too_big);
             column.insert_string(ndx, value.get_string());
             break;
         case type_Binary:
+            if (TIGHTDB_UNLIKELY(value.get_binary().size() > max_binary_size))
+                throw LogicError(LogicError::binary_too_big);
             column.insert_binary(ndx, value.get_binary());
             break;
         case type_Table: